4.rd/1.IAChronicleDetector/0.IAChronicleDetection/3.MachineLearningTranscription-Hybrid/pyproject.toml
[project]
name = "3-machinelearningtranscription-hybrid"
version = "0.1.0"
description = "Détection de chroniques radio via CamemBERT + Bi-LSTM + CRF"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=0.27.0",
    "joblib>=1.3.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "pytorch-crf>=0.7.2",
    "scikit-learn>=1.4.0",
    "sentencepiece>=0.2.0",
    "srt>=3.5.3",
    "torch>=2.2.0",
    "tqdm>=4.66.0",
    "transformers<5.0.0,>=4.38.0",
    "huggingface-hub>=0.20.0",
]
